migrations/models/4_20250221133645_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "bybitsecretkey" ADD "expires_at" TIMESTAMPTZ;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "bybitsecretkey" DROP COLUMN "expires_at";"""
