pyproject.toml
[project]
name = "bybit-info-telegram-bot"
version = "0.1.0"
description = ""
authors = [
    { name = "XetPy1030", email = "xetpy1030@mail.ru" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "aiogram (>=3.18.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pybit (>=5.9.0,<6.0.0)",
    "tortoise-orm[asyncpg] (>=0.24.0,<0.25.0)",
    "aerich (>=0.8.1,<0.9.0)",
    "redis (>=5.2.1,<6.0.0)",
    "structlog (>=25.1.0,<26.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]


[tool.aerich]
tortoise_orm = "app.database.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
